Added movement and basic animations
Added a Slingshot Goblin type. Added AnimBPs for Aura and enemies. Added a base BP_Enemy for the Goblin BPs to inherit from.
Added movement via wasd for Aura using the EnhancedInput system. Made Aura's camera stay in one position.
Added an interface for actors that can be moused over. In hindsight I should have called it "Hoverable". Allowed Enemies to be highlighted onmouseover to illustrate which one is "selected"